Make config file parsing logic more robust
The code for parsing the config file was quite naive; it expected the
file to have a very rigid structure, with no changes to whitespace etc.
It also contained a lot of copy-paste if-else statements.

This commit refactors config file parsing to be perform some more robust
text parsing when reading the file, while also replacing if-else chains
by loops that iterate over helper arrays.